fix: resolve repository and type annotation issues
- Added explicit type annotations for features variables in statistical_analysis.py
- Fixed failure_analysis.py return type by converting all values to float
- Improved InfluxDB manager mock class assignments to avoid type conflicts
- Fixed SQLAlchemy Column assignment issues by removing cast() and using type: ignore
- Resolved causal_relationship_repository set() argument type by extracting event IDs
- Added proper type annotations for dashboard_data and panel_data dictionaries
- Removed unused cast imports from repository modules

Progress: Reduced from 339 to 325 mypy errors (~4% improvement)

+++ escai_framework/analytics/failure_analysis.py
@@ -313,24 +313,24 @@ class RootCauseAnalyzer:
         """Extract features for causal analysis."""
         features = [
             # Sequence features
-            len(sequence.steps),
-            sequence.total_duration_ms,
-            sequence.success_rate,
-            sum(1 for step in sequence.steps if step.error_message),
+            float(len(sequence.steps)),
+            float(sequence.total_duration_ms),
+            float(sequence.success_rate),
+            float(sum(1 for step in sequence.steps if step.error_message)),
             
             # Epistemic state features
-            epistemic_state.confidence_level,
-            epistemic_state.uncertainty_score,
-            len(epistemic_state.belief_states),
-            len(epistemic_state.goal_states) if epistemic_state.goal_states else 0,
+            float(epistemic_state.confidence_level),
+            float(epistemic_state.uncertainty_score),
+            float(len(epistemic_state.belief_states)),
+            float(len(epistemic_state.goal_states) if epistemic_state.goal_states else 0),
             
             # Timing features
-            np.mean([step.duration for step in sequence.steps]) if sequence.steps else 0,
-            np.max([step.duration for step in sequence.steps]) if sequence.steps else 0,
+            float(np.mean([step.duration for step in sequence.steps]) if sequence.steps else 0),
+            float(np.max([step.duration for step in sequence.steps]) if sequence.steps else 0),
             
             # Complexity features
-            len(set(step.step_type for step in sequence.steps)),
-            np.mean([step.success_probability for step in sequence.steps]) if sequence.steps else 0
+            float(len(set(step.step_type for step in sequence.steps))),
+            float(np.mean([step.success_probability for step in sequence.steps]) if sequence.steps else 0)
         ]
         
         return features
